requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
jdcal==1.4.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.4
openpyxl==2.6.3
pandas==0.25.1
Pillow==6.1.0
protobuf==3.9.1
pyparsing==2.4.2
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.6
wget==3.2
wrapt==1.11.2
xlrd==1.2.0
